# openapi.yaml
openapi: 3.1.0
info:
  title: HTMLTrust Directory API
  description: |
    Reference implementation of the HTMLTrust trust directory API. Provides endpoints
    for author management, content signing and verification, claims, and a federated
    trust directory with reputation tracking.
  version: 1.0.0
  contact:
    name: HTMLTrust
    url: https://github.com/ArcadeLabsInc/htmltrust-server-reference
    email: jason@jason-grey.com
servers:
  - url: https://api.contentsigning.example.com/v1
    description: Production server
  - url: https://api.staging.contentsigning.example.com/v1
    description: Staging server
...
security:
  - {} # No default security
components:
  securitySchemes:
    AuthorApiKey:
      type: apiKey
      in: header
      name: X-AUTHOR-API-KEY
      description: API key linked to a specific author for author-specific operations

    GeneralApiKey:
      type: apiKey
      in: header
      name: X-API-KEY
      description: General API key for authenticated operations

    AdminApiKey:
      type: apiKey
      in: header
      name: X-ADMIN-API-KEY
      description: Admin-level API key for administrative operations
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
      example:
        code: "UNAUTHORIZED"
        message: "Invalid API key"
    Author:
      type: object
      required:
        - id
        - name
        - keyType
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the author
        name:
          type: string
          description: Name of the author
        description:
          type: string
          description: Description of the author
        url:
          type: string
          format: uri
          description: URL associated with the author
        keyType:
          type: string
          enum: [HUMAN, AI, HUMAN_AI_MIX, ORGANIZATION]
          description: Type of the author key
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        name: "John Doe"
        description: "Technology writer and researcher"
        url: "https://johndoe.example.com"
        keyType: "HUMAN"
        createdAt: "2023-01-01T12:00:00Z"
        updatedAt: "2023-01-01T12:00:00Z"
    PublicKey:
      type: object
      required:
        - id
        - authorId
        - key
        - algorithm
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the public key
        authorId:
          type: string
          format: uuid
          description: ID of the author this key belongs to
        key:
          type: string
          description: The public key in PEM format
        algorithm:
          type: string
          description: The cryptographic algorithm used
          enum: [RSA, ECDSA, ED25519]
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        expiresAt:
          type: string
          format: date-time
          description: Expiration timestamp (if applicable)
      example:
        id: "123e4567-e89b-12d3-a456-426614174001"
        authorId: "123e4567-e89b-12d3-a456-426614174000"
        key: "-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA...\n-----END PUBLIC KEY-----"
        algorithm: "RSA"
        createdAt: "2023-01-01T12:00:00Z"
        expiresAt: "2024-01-01T12:00:00Z"
...
    ContentSignature:
      type: object
      required:
        - contentHash
        - domain
        - authorId
        - signature
        - claims
      properties:
        contentHash:
          type: string
          description: Hash of the normalized content
        domain:
          type: string
          description: Domain associated with the content
        authorId:
          type: string
          format: uuid
          description: ID of the author who signed the content
        signature:
          type: string
          description: Cryptographic signature binding content, hash, domain, and author key
        claims:
          type: object
          description: Claims about the content
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
      example:
        contentHash: "sha256:a591a6d40bf420404a011733cfb7b190d62c65bf0bcda32b57b277d9ad9f146e"
        domain: "example.com"
        authorId: "123e4567-e89b-12d3-a456-426614174000"
        signature: "MEUCIQD7y5SxmQJ9f0lE9B0BwqIJKKdL5fZMNQOiPnKWUJfmrgIgEbHtPwDxM9xGbCZzW9k2R9jFxwJZQQlPfhgj+0YP7vQ="
        claims:
          ContentType: "Article"
          License: "CC-BY-4.0"
          AIAssistance: "Minimal"
        createdAt: "2023-01-01T12:00:00Z"
    KeyReputation:
      type: object
      required:
        - keyId
        - trustScore
        - verifiedSignatures
      properties:
        keyId:
          type: string
          format: uuid
          description: ID of the public key
        trustScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Trust score between 0 and 1
        verifiedSignatures:
          type: integer
          description: Number of verified signatures using this key
        reports:
          type: integer
          description: Number of reports against this key
        lastUpdated:
          type: string
          format: date-time
          description: Last update timestamp
      example:
        keyId: "123e4567-e89b-12d3-a456-426614174001"
        trustScore: 0.95
        verifiedSignatures: 1250
        reports: 2
        lastUpdated: "2023-01-01T12:00:00Z"
    ContentOccurrence:
      type: object
      required:
        - url
        - domain
        - firstSeen
      properties:
        url:
          type: string
          format: uri
          description: URL where the content was found
        domain:
          type: string
          description: Domain where the content was found
        firstSeen:
          type: string
          format: date-time
          description: When the content was first seen at this location
        lastSeen:
          type: string
          format: date-time
          description: When the content was last seen at this location
        authorId:
          type: string
          format: uuid
          description: ID of the author who signed the content at this location (if any)
        signatureValid:
          type: boolean
          description: Whether the signature at this location is valid
      example:
        url: "https://example.com/article/123"
        domain: "example.com"
        firstSeen: "2023-01-01T12:00:00Z"
        lastSeen: "2023-01-02T12:00:00Z"
        authorId: "123e4567-e89b-12d3-a456-426614174000"
        signatureValid: true
paths:
...
  /directory/keys:
    get:
      tags:
        - Directory
      summary: Search public keys
      description: Searches for public keys in the directory
      operationId: searchPublicKeys
      parameters:
        - name: authorName
          in: query
          description: Filter by author name
          schema:
            type: string
        - name: keyType
          in: query
          description: Filter by key type
          schema:
            type: string
            enum: [HUMAN, AI, HUMAN_AI_MIX, ORGANIZATION]
        - name: domain
          in: query
          description: Filter by domain associated with the author
          schema:
            type: string
        - name: minTrustScore
          in: query
          description: Minimum trust score
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 1
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: List of public keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/PublicKey"
                        - type: object
                          properties:
                            author:
                              $ref: "#/components/schemas/Author"
                            trustScore:
                              type: number
                              format: float
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      pages:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
  /directory/keys/{keyId}/reputation:
    get:
      tags:
        - Directory
      summary: Get key reputation
      description: Returns reputation information for a specific key
      operationId: getKeyReputation
      parameters:
        - name: keyId
          in: path
          required: true
          description: ID of the public key
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Key reputation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyReputation"
        "404":
          description: Key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /directory/keys/{keyId}/report:
    post:
      tags:
        - Directory
      summary: Report a key
      description: Reports a key for suspicious or malicious activity
      operationId: reportKey
      security:
        - GeneralApiKey: []
      parameters:
        - name: keyId
          in: path
          required: true
          description: ID of the public key
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  enum: [IMPERSONATION, MISINFORMATION, SPAM, OTHER]
                  description: Reason for reporting
                details:
                  type: string
                  description: Additional details about the report
                evidence:
                  type: string
                  format: uri
                  description: URL to evidence supporting the report
      responses:
        "201":
          description: Report submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  reportId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [PENDING, UNDER_REVIEW, ACCEPTED, REJECTED]
                    default: PENDING
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /directory/content:
    get:
      tags:
        - Directory
      summary: Search signed content
      description: Searches for signed content in the directory
      operationId: searchSignedContent
      parameters:
        - name: contentHash
          in: query
          description: Filter by content hash
          schema:
            type: string
        - name: authorId
          in: query
          description: Filter by author ID
          schema:
            type: string
            format: uuid
        - name: domain
          in: query
          description: Filter by domain
          schema:
            type: string
        - name: claim
          in: query
          description: "Filter by claim (format: claimName:claimValue)"
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: List of signed content
          content:
            application/json:
              schema:
                type: object
                properties:
                  signatures:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/ContentSignature"
                        - type: object
                          properties:
                            author:
                              $ref: "#/components/schemas/Author"
                            occurrences:
                              type: integer
                              description: Number of occurrences of this content
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      pages:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
  /directory/content/{contentHash}/occurrences:
    get:
      tags:
        - Directory
      summary: Find content occurrences
      description: Returns occurrences of a specific content hash across domains
      operationId: findContentOccurrences
      parameters:
        - name: contentHash
          in: path
          required: true
          description: Hash of the content
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: List of content occurrences
          content:
            application/json:
              schema:
                type: object
                properties:
                  occurrences:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContentOccurrence"
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      pages:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
        "404":
          description: Content hash not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /directory/content/report:
    post:
      tags:
        - Directory
      summary: Report content misuse
      description: Reports content for misuse, theft, or other issues
      operationId: reportContentMisuse
      security:
        - GeneralApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contentHash
                - sourceUrl
                - targetUrl
                - reason
              properties:
                contentHash:
                  type: string
                  description: Hash of the content being reported
                sourceUrl:
                  type: string
                  format: uri
                  description: Original source URL of the content
                targetUrl:
                  type: string
                  format: uri
                  description: URL where the content is being misused
                reason:
                  type: string
                  enum:
                    [
                      COPYRIGHT_VIOLATION,
                      UNAUTHORIZED_USE,
                      IMPERSONATION,
                      OTHER,
                    ]
                  description: Reason for reporting
                details:
                  type: string
                  description: Additional details about the report
      responses:
        "201":
          description: Report submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  reportId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [PENDING, UNDER_REVIEW, ACCEPTED, REJECTED]
                    default: PENDING
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Content hash not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"